# docs/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "The Electrum Bill Payments Service describes an interface for performing\
    \ bill payments\n                  transactions. These transactions require a\
    \ reference number which uniquely identifies the bill\n                  processor,\
    \ bill issuer, as well as the customer account, and is typically printed on a\
    \ customer\n                  invoice. The customer then makes use of this reference\
    \ number to make a payment towards the\n                  corresponding bill."
  version: "1.0.0"
  title: "Bill Payments Service Interface"
  contact:
    name: "Electrum Support"
    url: "http://io.electrum"
    email: "support@electrum.co.za"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "sandbox.electrum.io"
basePath: "/billpay/v1"
schemes:
- "https"
paths:
...
  /payments:
    post:
      summary: "Initiate a bill payment transaction"
      description: "Request that a payment be made towards a customer account"
      operationId: "createPayment"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "A payment request"
        required: true
        schema:
          $ref: "#/definitions/PaymentRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/PaymentResponse"
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/PaymentResponse"
          headers:
            Location:
              type: "string"
              description: "The location of the created payments resource"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/ErrorDetail"
        500:
          description: "Internal Server Error - MUST REVERSE"
          schema:
            $ref: "#/definitions/ErrorDetail"
        503:
          description: "Service Unavailable"
          schema:
            $ref: "#/definitions/ErrorDetail"
        504:
          description: "Gateway Timeout - MUST REVERSE"
          schema:
            $ref: "#/definitions/ErrorDetail"
      security:
      - httpBasic: []
  /payments/refunds:
    post:
      summary: "Creates a refund of previously confirmed payment"
      description: "If a payment is completed and confirmed successfully, some services\
        \ support that customers may request a refund for a particular payment for\
        \ some time after the payment took place. Not all services support refunds.\
        \ In the case where this function is not supported for the requested service,\
        \ a 501 HTTP status code may be returned"
      operationId: "createRefund"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "A refund request"
        required: true
        schema:
          $ref: "#/definitions/RefundRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/RefundResponse"
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/RefundResponse"
          headers:
            Location:
              type: "string"
              description: "The location of the created refund resource"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/ErrorDetail"
        500:
          description: "Internal Server Error - MUST REVERSE"
          schema:
            $ref: "#/definitions/ErrorDetail"
        501:
          description: "Not implemented - in the case where the service does not support\
            \ refunds"
          schema:
            $ref: "#/definitions/ErrorDetail"
        503:
          description: "Service Unavailable"
          schema:
            $ref: "#/definitions/ErrorDetail"
        504:
          description: "Gateway Timeout - MUST REVERSE"
          schema:
            $ref: "#/definitions/ErrorDetail"
      security:
      - httpBasic: []
  /payments/refunds/reversals:
    post:
      summary: "Reverse a refund request that failed or timed out"
      description: "If a createRefund request fails with a 500 or 504 HTTP status\
        \ code, or no response was received within the timeout period, it must be\
        \ reversed to ensure the payment is not refelected on a customer's account.\
        \ reverseRefund must be repeated until a final HTTP status code is received\
        \ (not 500 or 504). reverseRefund may be called repeatedly on the same payment\
        \ resource without negative effect."
      operationId: "reverseRefund"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "The RefundRequest originally sent in the createRefund request"
        required: true
        schema:
          $ref: "#/definitions/RefundRequest"
      responses:
        202:
          description: "Accepted"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/ErrorDetail"
        500:
          description: "Internal Server Error - MUST REVERSE"
          schema:
            $ref: "#/definitions/ErrorDetail"
        503:
          description: "Service Unavailable"
          schema:
            $ref: "#/definitions/ErrorDetail"
        504:
          description: "Gateway Timeout - MUST REVERSE"
          schema:
            $ref: "#/definitions/ErrorDetail"
      security:
      - httpBasic: []
  /payments/refunds/{refundId}/cancellations:
    post:
      summary: "Cancel an existing bill payment refund"
      description: "If a createRefund request previously succeeded with a 201 status\
        \ it must be confirmed or cancelled to complete the transaction. cancelRefund\
        \ can only suceeed if a refund was created but not confirmed. cancelRefund\
        \ must be repeated until a final HTTP status code is received (not 500 or\
        \ 504). If a status code of either 500 or 504 is received, or no response\
        \ is received, the request must be repeated. cancelRefund may be called repeatedly\
        \ on the same payment resource without negative effect."
      operationId: "cancelRefund"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "refundId"
        in: "path"
        description: "The id returned in the corresponding createRefund response"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "A cancellation"
        required: true
        schema:
          $ref: "#/definitions/Cancellation"
      responses:
        202:
          description: "Accepted"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/ErrorDetail"
        500:
          description: "Internal Server Error - MUST REPEAT"
          schema:
            $ref: "#/definitions/ErrorDetail"
        503:
          description: "Service Unavailable"
          schema:
            $ref: "#/definitions/ErrorDetail"
        504:
          description: "Gateway Timeout - MUST REPEAT"
          schema:
            $ref: "#/definitions/ErrorDetail"
      security:
      - httpBasic: []
  /payments/refunds/{refundId}/confirmations:
    post:
      summary: "Confirm an existing bill payment refund"
      description: "If a createRefund request previously succeeded with a 201 status\
        \ it must be confirmed or cancelled to complete the transaction. confirmRefund\
        \ can only suceeed if a refund was created but not cancelled. confirmRefund\
        \ must be repeated until a final HTTP status code is received (not 500 or\
        \ 504). If a status code of either 500 or 504 is received, or no response\
        \ is received, the request must be repeated. confirmRefund may be called repeatedly\
        \ on the same refund resource without negative effect."
      operationId: "confirmRefund"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "refundId"
        in: "path"
        description: "The id returned in the corresponding createRefund response"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "A refund confirmation"
        required: true
        schema:
          $ref: "#/definitions/RefundConfirmation"
      responses:
        202:
          description: "Accepted"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/ErrorDetail"
        500:
          description: "Internal Server Error - MUST REPEAT"
          schema:
            $ref: "#/definitions/ErrorDetail"
        503:
          description: "Service Unavailable"
          schema:
            $ref: "#/definitions/ErrorDetail"
        504:
          description: "Gateway Timeout - MUST REPEAT"
          schema:
            $ref: "#/definitions/ErrorDetail"
      security:
      - httpBasic: []
  /payments/reversals:
    post:
      summary: "Reverse a payment request that failed or timed out"
      description: "If a createPayment request fails with a 500 or 504 HTTP status\
        \ code, or no response was received within the timeout period, it must be\
        \ reversed to ensure the payment is not refelected on a customer's account.\
        \ reversePayment must be repeated until a final HTTP status code is received\
        \ (not 500 or 504). reversePayment may be called repeatedly on the same payment\
        \ resource without negative effect."
      operationId: "reversePayment"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "The PaymentRequest originally sent in the createPayment request"
        required: true
        schema:
          $ref: "#/definitions/PaymentRequest"
      responses:
        202:
          description: "Accepted"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/ErrorDetail"
        500:
          description: "Internal Server Error - MUST REVERSE"
          schema:
            $ref: "#/definitions/ErrorDetail"
        503:
          description: "Service Unavailable"
          schema:
            $ref: "#/definitions/ErrorDetail"
        504:
          description: "Gateway Timeout - MUST REVERSE"
          schema:
            $ref: "#/definitions/ErrorDetail"
      security:
      - httpBasic: []
  /payments/{paymentId}/cancellations:
    post:
      summary: "Cancel an existing bill payment"
      description: "If a createPayment request previously succeeded with a 201 status\
        \ it must be confirmed or cancelled to complete the transaction. cancelPayment\
        \ can only suceeed if a payment was created but not confirmed. cancelPayment\
        \ must be repeated until a final HTTP status code is received (not 500 or\
        \ 504). If a status code of either 500 or 504 is received, or no response\
        \ is received, the request must be repeated. cancelPayment may be called repeatedly\
        \ on the same payment resource without negative effect."
      operationId: "cancelPayment"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "paymentId"
        in: "path"
        description: "The id returned in the corresponding createPayment response"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "A cancellation"
        required: true
        schema:
          $ref: "#/definitions/Cancellation"
      responses:
        202:
          description: "Accepted"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/ErrorDetail"
        500:
          description: "Internal Server Error - MUST REPEAT"
          schema:
            $ref: "#/definitions/ErrorDetail"
        503:
          description: "Service Unavailable"
          schema:
            $ref: "#/definitions/ErrorDetail"
        504:
          description: "Gateway Timeout - MUST REPEAT"
          schema:
            $ref: "#/definitions/ErrorDetail"
      security:
      - httpBasic: []
  /payments/{paymentId}/confirmations:
    post:
      summary: "Confirm an existing bill payment"
      description: "If a createPayment request previously succeeded with a 201 status\
        \ it must be confirmed or cancelled to complete the transaction. confirmPayment\
        \ can only suceeed if a payment was created but not cancelled. confirmPayment\
        \ must be repeated until a final HTTP status code is received (not 500 or\
        \ 504). If a status code of either 500 or 504 is received, or no response\
        \ is received, the request must be repeated. confirmPayment may be called\
        \ repeatedly on the same payment resource without negative effect."
      operationId: "confirmPayment"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "paymentId"
        in: "path"
        description: "The id returned in the corresponding createPayment response"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "A payment confirmation"
        required: true
        schema:
          $ref: "#/definitions/PaymentConfirmation"
      responses:
        202:
          description: "Accepted"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/ErrorDetail"
        500:
          description: "Internal Server Error - MUST REPEAT"
          schema:
            $ref: "#/definitions/ErrorDetail"
        503:
          description: "Service Unavailable"
          schema:
            $ref: "#/definitions/ErrorDetail"
        504:
          description: "Gateway Timeout - MUST REPEAT"
          schema:
            $ref: "#/definitions/ErrorDetail"
      security:
      - httpBasic: []
securityDefinitions:
  httpBasic:
    description: "All requests require HTTP basic authentication, where user name\
      \ equals the value used in\n                    MessageId.institutionId and\
      \ password has been agreed with Electrum."
    type: "basic"
definitions:
  Account:
    type: "object"
    required:
    - "accountRef"
    properties:
      accountRef:
        type: "string"
        description: "A reference number identifying the bill payments processor,\
          \ bill issuer, and customer"
        minLength: 6
        maxLength: 40
      balance:
        $ref: "#/definitions/LedgerAmount"
      dueDate:
        type: "string"
        format: "date-time"
        description: "The effective date of the current billing period"
    description: "Represents the status of a customer account"
...
  Cancellation:
    type: "object"
    required:
    - "messageId"
    properties:
      linkData:
        type: "object"
        description: "The unaltered linkData object as supplied in the createPayment\
          \ response. Required if the createPayment response contained linkData"
      messageId:
        $ref: "#/definitions/MessageId"
    description: "Represents the data required to cancel a previously created payment"
  Customer:
    type: "object"
    properties:
      firstName:
        type: "string"
        description: "The customer's first name(s)"
        minLength: 0
        maxLength: 40
      lastName:
        type: "string"
        description: "The customer's last name"
        minLength: 0
        maxLength: 40
      address:
        type: "string"
        description: "The customer's address"
        minLength: 0
        maxLength: 80
    description: "A bill payment customer"
  ErrorDetail:
    type: "object"
    required:
    - "responseMessage"
    properties:
      responseCode:
        type: "string"
        description: "The response code. Will always be present unless request to\
          \ upstream timed out"
      responseMessage:
        type: "string"
        description: "The response code description"
      detailMessage:
        type: "object"
        description: "A free form detailed description of a particular failure condition\
          \ may optionally be supplied"
    description: "Represents the outcome of a completed transaction"
  Institution:
    type: "object"
    required:
    - "id"
    - "name"
    properties:
      id:
        type: "string"
        description: "The institution's id"
        pattern: "[0-9]{1,11}"
      name:
        type: "string"
        description: "The institutions's name"
        minLength: 0
        maxLength: 40
    description: "Processing or receiving institution details"
...
  Merchant:
    type: "object"
    required:
    - "merchantId"
    - "merchantName"
    - "merchantType"
    properties:
      merchantType:
        type: "string"
        description: "The assigned four digit merchant category code"
        pattern: "[0-9]{4}"
      merchantId:
        type: "string"
        description: "The assigned merchant identifier. Also known as card acceptor\
          \ id"
        minLength: 15
        maxLength: 15
      merchantName:
        $ref: "#/definitions/MerchantName"
    description: "Merchant related data. Must be included if available"
...
  PaymentConfirmation:
    type: "object"
    required:
    - "messageId"
    - "tenders"
    properties:
      linkData:
        type: "object"
        description: "The unaltered linkData object as supplied in the createPayment\
          \ response. Required if the createPayment response contained linkData"
      messageId:
        $ref: "#/definitions/MessageId"
      tenders:
        type: "array"
        description: "An array of tenders used to pay for the transaction"
        items:
          $ref: "#/definitions/Tender"
        minItems: 1
    description: "Represents the data required to confirm a previously created payment"
  PaymentRequest:
    type: "object"
    required:
    - "accountRef"
    - "messageId"
    - "requestDetail"
    properties:
      accountRef:
        type: "string"
        description: "A reference number identifying the bill payments processor,\
          \ bill issuer, and customer"
        minLength: 6
        maxLength: 40
      messageId:
        $ref: "#/definitions/MessageId"
      merchant:
        $ref: "#/definitions/Merchant"
      requestDetail:
        $ref: "#/definitions/RequestDetail"
    description: "Represents a request to perform a payment"
  PaymentResponse:
    type: "object"
    required:
    - "account"
    - "id"
    - "receiver"
    - "responseDetail"
    - "slipData"
    properties:
      id:
        type: "string"
        description: "An identifier that uniquely identifies the payment"
        pattern: "[A-Za-z0-9._~-]{1,40}"
      account:
        $ref: "#/definitions/Account"
      customer:
        $ref: "#/definitions/Customer"
      linkData:
        type: "object"
        description: "This object may be used by the server to link a payment to a\
          \ payment confirmation or cancellation, and should be provided unaltered\
          \ by clients in corresponding requests"
      processor:
        $ref: "#/definitions/Institution"
      receiver:
        $ref: "#/definitions/Institution"
      responseDetail:
        $ref: "#/definitions/ResponseDetail"
      slipData:
        $ref: "#/definitions/SlipData"
    description: "Represents a response to a payment request"
  RefundConfirmation:
    type: "object"
    required:
    - "messageId"
    properties:
      linkData:
        type: "object"
        description: "The unaltered linkData object as supplied in the createRefund\
          \ response. Required if the createRefund response contained linkData"
      messageId:
        $ref: "#/definitions/MessageId"
    description: "Represents the data required to confirm a previously created refund"
  RefundRequest:
    type: "object"
    required:
    - "issuerReference"
    - "messageId"
    - "refundReason"
    properties:
      messageId:
        $ref: "#/definitions/MessageId"
      issuerReference:
        type: "string"
        description: "An identifier that is printed on the customer slip and uniquely\
          \ identifies the payment on the bill issuer's system. This value is used\
          \ by the customer to request a refund when the service supports this function,\
          \ and it is thus important that this number is unique"
        pattern: "[A-Z0-9]{1,20}"
      refundReason:
        type: "string"
        description: "The reason for the refund as given by the customer"
        minLength: 0
        maxLength: 20
    description: "Represents a request to refund a payment"
  RefundResponse:
    type: "object"
    required:
    - "account"
    - "id"
    - "receiver"
    properties:
      id:
        type: "string"
        description: "An identifier that uniquely identifies the payment"
        pattern: "[A-Za-z0-9._~-]{1,40}"
      account:
        $ref: "#/definitions/Account"
      customer:
        $ref: "#/definitions/Customer"
      linkData:
        type: "object"
        description: "This object may be used by the server to link a refund to a\
          \ refund confirmation or cancellation, and should be provided unaltered\
          \ by clients in corresponding requests"
      processor:
        $ref: "#/definitions/Institution"
      receiver:
        $ref: "#/definitions/Institution"
    description: "Represents a response to a refund request"
  RequestDetail:
    type: "object"
    required:
    - "requestAmount"
    properties:
      clientRef:
        type: "string"
        description: "A reference number useful to the client for identifying transactions,\
          \ also knows as a retrieval reference number"
        pattern: "[A-Za-z0-9 ]{0,12}"
      requestAmount:
        $ref: "#/definitions/LedgerAmount"
    description: "Request specific data"
...